日期格式去掉特殊字符改为int类型

+++ src/main/scala/com/jeff/financing/entity/Stocktaking.scala
@@ -9,13 +9,13 @@ import reactivemongo.api.bson.BSONObjectID
  *
  * @param _id
  * @param targetId   资产id
- * @param date       盘点日期yyyy-MM-dd
+ * @param date       盘点日期yyyyMM,例如202012
  * @param amount     金额
  * @param comment    备注
  * @param createTime 创建时间
  */
 @Persistence(collName = "stocktaking")
-case class Stocktaking(_id: Option[BSONObjectID], targetId: String, date: Long, amount: BigDecimal, comment: Option[String], createTime: Long)
+case class Stocktaking(_id: Option[BSONObjectID], targetId: String, date: Int, amount: BigDecimal, comment: Option[String], createTime: Long)
 
 object StocktakingJsonSupport extends ObjectIdSerialization with LowerCaseWithUnderscores with SprayJsonSupport {
   implicit val stocktakingFormats = jsonFormat6(Stocktaking)
